fix(load-data): avoid circular import
Moves loadDataSources logic to composable,
keeps store mostly state.
